test: split service health tests
